+++ test/ERC20SnapshotModuleOwnable2Step.test.js
@@ -1,12 +1,6 @@
-const ERC20SnapshotModuleCommonRescheduling = require('./ERC20SnapshotModuleCommon/ERC20SnapshotModuleCommonRescheduling')
-const ERC20SnapshotModuleCommonScheduling = require('./ERC20SnapshotModuleCommon/ERC20SnapshotModuleCommonScheduling')
-const ERC20SnapshotModuleCommonUnschedule = require('./ERC20SnapshotModuleCommon/ERC20SnapshotModuleCommonUnschedule')
-const ERC20SnapshotModuleCommonGetNextSnapshot = require('./ERC20SnapshotModuleCommon/ERC20SnapshotModuleCommonGetNextSnapshot')
-const ERC20SnapshotModuleCommonExact = require('./ERC20SnapshotModuleCommon/ERC20SnapshotModuleCommonExact')
-const ERC20SnapshotModuleCommonMaterialized = require('./ERC20SnapshotModuleCommon/ERC20SnapshotModuleCommonMaterialized')
-const ERC20SnapshotModuleMultiplePlannedTest = require('./ERC20SnapshotModuleCommon/global/ERC20SnapshotModuleMultiplePlannedTest')
-const ERC20SnapshotModuleOnePlannedSnapshotTest = require('./ERC20SnapshotModuleCommon/global/ERC20SnapshotModuleOnePlannedSnapshotTest')
-const ERC20SnapshotModuleZeroPlannedSnapshotTest = require('./ERC20SnapshotModuleCommon/global/ERC20SnapshotModuleZeroPlannedSnapshot')
+const {
+  registerSnapshotCommonSuites
+} = require('./ERC20SnapshotModuleCommon/registerSnapshotCommonSuites')
 const {
   deployCMTATStandalone,
   fixture,
@@ -28,13 +22,5 @@ describe('Standard - ERC20SnapshotModule Ownable2Step', function () {
     this.cmtat.connect(this.admin).setSnapshotEngine(this.transferEngineMock)
   })
 
-  ERC20SnapshotModuleMultiplePlannedTest()
-  ERC20SnapshotModuleOnePlannedSnapshotTest()
-  ERC20SnapshotModuleZeroPlannedSnapshotTest()
-  ERC20SnapshotModuleCommonRescheduling()
-  ERC20SnapshotModuleCommonScheduling()
-  ERC20SnapshotModuleCommonUnschedule()
-  ERC20SnapshotModuleCommonGetNextSnapshot()
-  ERC20SnapshotModuleCommonExact()
-  ERC20SnapshotModuleCommonMaterialized()
+  registerSnapshotCommonSuites()
 })